+++ src/main/java/bowling/domain/FrameResult.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 import java.util.ArrayList;
 import java.util.List;
