+++ src/test/java/bowling/domain/NormalFrameTest.java
@@ -1,8 +1,10 @@
-package bowling;
+package bowling.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import bowling.domain.FrameResult;
+import bowling.domain.NormalFrame;
 import java.util.Arrays;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
